requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
geopy==2.3.0
httpx==0.24.0
python-dotenv==1.0.0
pillow==9.5.0
numpy==1.24.3
aiofiles==23.1.0